Neural Style Transfer/requirements.txt
Flask
torch
torchvision
Pillow
numpy
matplotlib
keras
tensorflow